extract patcher to own module
